Refactor processors of Android adaptive icons
