+++ lib/src/processors/android/icons/android_icons_processor.dart
@@ -25,7 +25,7 @@
 
 import 'package:flutter_flavorizr/src/extensions/extensions_map.dart';
 import 'package:flutter_flavorizr/src/parser/models/flavorizr.dart';
-import 'package:flutter_flavorizr/src/processors/android/icons/android_adaptive_icon_processor.dart';
+import 'package:flutter_flavorizr/src/processors/android/icons/android_adaptive_icons_processor.dart';
 import 'package:flutter_flavorizr/src/processors/android/icons/android_adaptive_icon_xml_processor.dart';
 import 'package:flutter_flavorizr/src/processors/android/icons/android_icon_processor.dart';
 import 'package:flutter_flavorizr/src/processors/commons/queue_processor.dart';
@@ -66,7 +66,7 @@ class AndroidIconsProcessor extends QueueProcessor {
                 .map(
                   (flavorName, flavor) => MapEntry(
                     flavorName,
-                    AndroidAdaptiveIconProcessor(
+                    AndroidAdaptiveIconsProcessor(
                       flavor.android!.adaptiveIcon!.foreground,
                       flavor.android!.adaptiveIcon!.background,
                       flavorName,
